+++ NEWS.md
@@ -3,7 +3,7 @@
 ## New features
 
 - Support for "predict_parts" objects from "DALEX" package.
-- "shapviz" objects `x1, x2` can now be concatenated using `x1 + x2`.
+- "shapviz" objects `x1, x2` can now be concatenated using `x1 + x2` or `rbind(x1, x2)`.
 
 # shapviz 0.6.1